+++ sickrage/providers/torrent/cpasbien.py
@@ -54,7 +54,7 @@ class CpasbienProvider(TorrentProvider):
                 sickrage.srCore.srLogger.debug("Search URL: %s" % searchURL)
 
                 try:
-                    data = sickrage.srCore.srWebSession.get(searchURL, cache=False).text
+                    data = sickrage.srCore.srWebSession.get(searchURL).text
                 except Exception:
                     sickrage.srCore.srLogger.debug("No data returned from provider")
                     continue