+++ sickrage/providers/torrent/morethantv.py
@@ -114,7 +114,7 @@ class MoreThanTVProvider(TorrentProvider):
 
                 # returns top 15 results by default, expandable in user profile to 100
                 try:
-                    data = sickrage.srCore.srWebSession.get(searchURL, cache=False).text
+                    data = sickrage.srCore.srWebSession.get(searchURL).text
                 except Exception:
                     sickrage.srCore.srLogger.debug("No data returned from provider")
                     continue