+++ changelog.md
@@ -1,6 +1,14 @@
 # Changelog
 
-- * 821dc0d - 2017-10-28: Release v9.1.60 
+- * 196db64 - 2017-10-29: Release v9.1.62 
+- * 8191cb2 - 2017-10-29: Fixed categories for TorrentBytes 
+- * f49f2ee - 2017-10-29: Added custom url setting to IPTorrents 
+- * 0c01ea0 - 2017-10-29: Cleaned up IPTorrents provider code 
+- * 598a329 - 2017-10-29: Fixed issus with TorrentBytes provider Fixed issues with TorrentLeech provider 
+- * 415189d - 2017-10-28: Fixed typo in TorrentBytes provider 
+- * 7b6bbe3 - 2017-10-28: Enabled caching for all providers Fixed issue #99 - Torrentbytes not returning search results 
+- * a90e20b - 2017-10-28: Fixed default poster size 
+- * ad2f5d9 - 2017-10-28: Release v9.1.60 
 - * b30936a - 2017-10-28: Release v9.1.59 
 - * 56a04dc - 2017-10-28: Fixed issue with verifying search results 
 - * 2c23053 - 2017-10-28: Fixed issues with scene numbering and NoneType errors